requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
elevenlabs==1.55.0
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
numpy==2.2.4
protobuf==5.29.4
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
requests==2.32.3
sniffio==1.3.1
sounddevice==0.5.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
websockets==15.0.1
python-dotenv==1.0.1
